fix: directory detection with remote executor
Previously, directories created during a Snakemake rule were initially
cached as non-directories due to their nascent state in storage. Using a
local executor, this would not be an issue because the cache would be
invalidated during upload. However, this is not the case with remote
executors. Therefore, Snakemake's would incorrectly report these
directories as non-existent, causing rule failure for no reason.

This commit resolves the problem by deferring the caching of file types
until their existence and type (file or directory) are definitive. This
ensures that the plugin always returns a definitive state.